gst/msdk: move HEVC encoder test classes to lib
